requirements.txt
PyQt6
Pillow
pytest
ruff
black
rawpy
